moved transaction tests into connected folder
