+++ src/app/components/blockchain/Connected/TestSendTransaction.tsx
@@ -1,5 +1,5 @@
 import { useEffect, useState } from "react";
-import CardButton from "./CardButton/CardButton";
+import CardButton from "../../CardButton/CardButton";
 import { useSendTransaction, useWalletClient } from "wagmi";
 import React from "react";
 import { Box, Text } from "@0xsequence/design-system";
